requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
yfinance==0.2.31
pandas==2.1.1
newsapi-python==0.2.7
python-dotenv==1.0.0
streamlit==1.32.0
plotly==5.18.0
matplotlib==3.8.2
scipy==1.12.0
